+++ datagouv/downloader/_resources_downloader__test.py
@@ -3,10 +3,20 @@ import pytest
 from datagouv import ResourcesDownloader
 from datagouv.downloader._resources_downloader import RESOURCE_TYPES
 
-resource_1 = {"id": "r_1_id", "url": "r_1_url", "type": "main"}
-resource_2 = {"id": "r_2_id", "url": "r_2_url_2024", "type": "main"}
-resource_3 = {"id": "r_3_id", "url": "r_3_url_2024", "type": "documentation"}
-resource_4 = {"id": "r_4_id", "url": "r_4_url", "type": "other"}
+resource_1 = {"id": "r_1_id", "url": "r_1_url", "title": "1 Url", "type": "main"}
+resource_2 = {
+    "id": "r_2_id",
+    "url": "r_2_url_2024",
+    "title": "2 url_2024",
+    "type": "main",
+}
+resource_3 = {
+    "id": "r_3_id",
+    "url": "r_3_url_2024",
+    "title": "3 url_2024",
+    "type": "documentation",
+}
+resource_4 = {"id": "r_4_id", "url": "r_4_url", "title": "4 Url", "type": "other"}
 
 dataset_1_id = "id_d_1"
 dataset_1 = {
@@ -73,9 +83,24 @@ def test_main_and_doc_files():
     assert len(d.urls) == 3
 
 
-def test_regex_urls():
+def test_regex_titles():
+    title_regex = "_2024"
+    types = ["main"]
     d = ResourcesDownloader(
-        dataset_1_id, ["main"], "_2024", DatagouvClient=MockDatagouvClient
+        dataset_1_id, types, title_regex, DatagouvClient=MockDatagouvClient
+    )
+    # Test files types
+    assert len(d.resources) == 4
+
+    # Only one url is type=main and contains '_2024'
+    assert len(d.urls) == 1
+
+
+def test_advanced_regex_titles():
+    title_regex = "_[0-9]{4}"
+    types = ["main"]
+    d = ResourcesDownloader(
+        dataset_1_id, types, title_regex, DatagouvClient=MockDatagouvClient
     )
     # Test files types
     assert len(d.resources) == 4